tests/requirements.txt
selenium==4.15.0
webdriver-manager==4.0.1
pytest==8.3.3
pytest-html==4.1.1
pytest-xdist==3.3.1
python-dotenv==1.0.0